Move GetReadVersion* structs to GrvProxyInterface.h
